refactoring and renaming
This renames the extractor package such that it represents the type it works with. The aim of this is to keep a clean API within the library.

When other types are added they will have a package named accordingly and then the usage just reads type.Extract which is really neat.

Tests will be added next, along with improved output from the CLI